+++ AdminPanel/client/src/pages/AiIntegration/ai/scripts/code.js
@@ -38,13 +38,27 @@ let summarizationWindow = null;
 let translateSettingsWindow = null;
 let helperWindow = null;
 
-window.addSupportAgentMode = function() {
+window.getActionsInfo = function() {
+	let actions = [];
+	for (const action in AI.ActionType) {
+		if (AI.ActionType.hasOwnProperty(action)) {
+			let requestEngine = AI.Request.create(action, true);
+			if (requestEngine)
+				actions.push({ [action] : requestEngine.model });
+		}		
+	}
+	return actions;
+};
+
+window.addSupportAgentMode = function(editorVersion) {
 	var agentHistory = [];
 	var agentDebug = false;
 
 	if (!window.EditorHelper) {
 		window.EditorHelper = new EditorHelperImpl();
 	}
+
+	var is91 = editorVersion >= 9001000;
 
 	window.Asc.plugin.attachEditorEvent("onKeyDown", function(e) {
 		if (e.keyCode === 27 && helperWindow) {
@@ -56,9 +70,21 @@ window.addSupportAgentMode = function() {
 
 		let isCtrl = e.ctrlKey || e.metaKey;
 		let isClearHistory = isCtrl && e.altKey;
-		let codeShow = 191; // '/'
 
-		if (e.keyCode === codeShow && isCtrl && !helperWindow) {
+		let isAgentShow = false;
+		if (isCtrl)
+		{
+			if (is91)
+			{
+				isAgentShow = e.key === "/";
+			}
+			else
+			{
+				isAgentShow = e.keyCode === 191 || e.keyCode === 111;
+			}
+		}		
+		
+		if (isAgentShow && isCtrl && !helperWindow) {
 			if (isClearHistory)
 				agentHistory = [];
 
@@ -296,30 +322,97 @@ async function initWithTranslate(counter) {
 		if (editorVersion >= 9000000) {
 			window.Asc.plugin.attachEditorEvent("onAIRequest", async function(params){
 				let data = {};
-				switch (params.type) {
-					case "text":
-					{
-						let requestEngine = AI.Request.create(AI.ActionType.Chat);
-						if (requestEngine)
-						{
-							let result = await requestEngine.chatRequest(params.data);
-							if (!result) result = "";
+				let isFromMethod = params.isFromMethod === true;
+				let isBlock = !isFromMethod;
 
-							data.type = "text";
-							data.text = result;
+				async function sendResult(data) {
+					if (isFromMethod) {
+						await Asc.Editor.callMethod("SendEventInternal", ["ai_onRequest", data]);	
+					} else {
+						await Asc.Editor.callMethod("onAIRequest", [data]);
+					}
+				}
+
+				if ("Actions" === params.type) {
+					data.Actions = window.getActionsInfo();
+					return await sendResult(data);
+				}
+				
+				if ("text" === params.type)
+					params.type = AI.ActionType.Chat;
+
+				let requestEngine = null;
+				if (AI.Actions[params.type])
+					requestEngine = AI.Request.create(params.type);
+
+				if (!requestEngine) {
+					data.type = "no-engine";
+					data.text = "";
+					data.error = "No model selected for chat action...";
+
+					return await sendResult(data);
+				}
+
+				if (isFromMethod) {
+					await Asc.Editor.callMethod("SendEventInternal", ["ai_onStartAction", {
+						type : "Block",
+						description : "AI (" + requestEngine.modelUI.name + ")"
+					}]);
+				}
+
+				switch (params.type) {
+					case AI.ActionType.Chat:
+					{
+						data.type = "text";
+						data.text = await requestEngine.chatRequest(params.data, isBlock);
+						break;
+					}
+					case AI.ActionType.Translation:
+					{
+						data.type = "text";
+						let prompt = Asc.Prompts.getTranslatePrompt(params.data.text, params.data.lang);
+						let result = await requestEngine.chatRequest(prompt, isBlock);
+						data.text = result ? Asc.Library.getTranslateResult(result, params.data.text) : "";
+						break;
+					}
+					case AI.ActionType.ImageGeneration:
+					{
+						data.type = AI.ActionType.ImageGeneration;
+						let result = await requestEngine.imageGenerationRequest(params.data, isBlock);
+						data.image = result || "";
+						break;
+					}
+					case AI.ActionType.OCR:
+					{
+						data.type = AI.ActionType.OCR;
+						let result = await requestEngine.imageOCRRequest(params.data, isBlock);
+						if (result) {
+							data.result = Asc.Library.ConvertMdToHTML(result, [Asc.PluginsMD.latex]);
 						}
-						else
-						{
-							data.type = "no-engine";
-							data.text = "";
-							data.error = "No model selected for chat action..."
-						}
+						break;
+					}
+					case AI.ActionType.Vision:
+					{
+						data.type = AI.ActionType.Vision;
+						let result = await requestEngine.imageVisionRequest({
+							prompt : Asc.Prompts.getImageDescription(),
+							image : params.data
+						}, isBlock);
+						data.result = result || "";
+						break;
 					}
 					default:
 						break;
 				}
 
-				await Asc.Editor.callMethod("onAIRequest", [data]);
+				if (isFromMethod) {
+					await Asc.Editor.callMethod("SendEventInternal", ["ai_onEndAction", {
+						type : "Block",
+						description : "AI (" + requestEngine.modelUI.name + ")"
+					}]);
+				}
+
+				await sendResult(data);			
 			});
 
 			if ("cell" === window.Asc.plugin.info.editorType) {
@@ -392,7 +485,7 @@ async function initWithTranslate(counter) {
 		}
 
 		if (editorVersion >= 9000004)
-			window.addSupportAgentMode();
+			window.addSupportAgentMode(editorVersion);
 	}
 }
 
@@ -430,6 +523,102 @@ window.Asc.plugin.init = async function() {
 			AI.serverSettings = null;
 		}
 		delete window.Asc.plugin.info.aiPluginSettings;
+	}
+
+	if (this.sendEvent) {
+		this.sendEvent("ai_onInit", {});
+		this.attachEditorEvent("ai_onCustomProviders", function(providers) {
+			
+			for (let i = 0, len = providers.length; i < len; i++) {
+				let item = providers[i];
+				if (!item.name)
+					continue;
+
+				if (!item.content) {
+					item.content = "\"use strict\";\n\
+class Provider extends AI.Provider {\n\
+	constructor() {\n\
+		super(\"" + item.name + "\", \"[external]\", \"\", \"\");\n\
+	}\n\
+}";
+					let isError = !AI.addCustomProvider(item.content);
+
+					if (!isError) {
+						customProvidersWindow && customProvidersWindow.command('onSetCustomProvider', AI.getCustomProviders());
+						aiModelEditWindow && aiModelEditWindow.command('onProvidersUpdate', { providers : AI.serializeProviders() });
+					}					
+				}
+			}
+
+		});
+
+		this.attachEditorEvent("ai_onCustomInit", function(obj) {
+			
+			if (obj.settingsLock !== undefined) {
+				let isSettingsRemoved = obj.settingsLock === "removed";
+				let isSettingsDisabled = obj.settingsLock === "disabled";
+
+				if (window.buttonSettings) {
+					if (window.buttonSettings.removed != isSettingsRemoved || 
+						window.buttonSettings.disabled != isSettingsDisabled) {
+						window.buttonSettings.removed = isSettingsRemoved;
+						window.buttonSettings.disabled = isSettingsDisabled;
+
+						Asc.Buttons.updateToolbarMenu(window.buttonMainToolbar.id, window.buttonMainToolbar.name, [window.buttonSettings]);
+					}
+				}
+			}
+
+			if (obj.actions) {
+				let isActionsOverride = obj.actionsOverride === true;
+
+				for (let type in obj.actions) {
+					if (!AI.Actions[type])
+						continue;
+					
+					if (!AI.Actions[type].model || isActionsOverride)
+						AI.Actions[type].model = obj.actions[type].model;
+				}
+
+				AI.ActionsSave();
+			}
+
+			let isUpdate = false;
+			if (obj.providers) {
+				for (let type in obj.providers) {
+					AI.Providers[type] = obj.providers[type];
+				}
+				isUpdate = true;
+			}
+
+			if (obj.models) {
+				for (let i = 0, len = obj.models.length; i < len; i++) {
+					let model = obj.models[i];
+					let isFound = false;
+
+					for (let j = 0, jLen = AI.Models.length; j < jLen; j++) {
+						let testModel = AI.Models[j];
+
+						if (testModel.name === model.name && 
+							testModel.provider === model.provider &&
+							testModel.id === model.id) {
+							isFound = true;
+							AI.Models[j] = model;
+							break;
+						}
+					}
+
+					if (!isFound) {
+						AI.Models.push(model);
+					}
+				}
+				isUpdate = true;
+			}
+
+			if (isUpdate)
+				AI.Storage.save();
+			
+		});
 	}
 
 	await initWithTranslate(1 << 1);
@@ -622,7 +811,7 @@ function onOpenEditModal(data) {
 		],
 		isModal : true,
 		EditorsSupport : ["word", "slide", "cell", "pdf"],
-		size : [320, 375]
+		size : [365, 425]
 	};
 
 	if (!aiModelEditWindow) {
@@ -642,6 +831,11 @@ function onOpenEditModal(data) {
 				model : data.model ? AI.Storage.getModelByName(data.model.name) : null,
 				providers : AI.serializeProviders()
 			});
+		});
+		aiModelEditWindow.attachEvent("onUpdateHeight", function(height) {
+			if(height > variation.size[1]) {
+				Asc.Editor.callMethod("ResizeWindow", [aiModelEditWindow.id, [variation.size[0] - 2, height]]);	//2 is the border-width at the window
+			}
 		});
 		aiModelEditWindow.attachEvent('onOpenCustomProvidersModal', onOpenCustomProvidersModal);
 	}