+++ AdminPanel/client/src/pages/AiIntegration/ai/scripts/helpers/word.js
@@ -37,11 +37,16 @@ var WORD_FUNCTIONS = {};
 	{
 		let func = new RegisteredFunction();
 		func.name = "commentText";
+		func.description = "Use this function if you asked to comment or explain anything. If text or paragraph number is not specified assume that we are working with the current paragraph. Specify whether the explanation should be added as a comment or as a footnote. The AI will generate the content based on your prompt and insert it in the chosen format.";
 		func.params = [
 			"type (string): whether to add as a 'comment' or as a 'footnote' (default is 'comment')"
 		];
 
+
 		func.examples = [
+			"If you need to explain something, respond with:\n" +
+			"[functionCalling (commentText)]: {\"prompt\" : \"Explain this text\", \"type\": \"comment\"}",
+
 			"If you need to explain selected text as a comment, respond with:\n" +
 			"[functionCalling (commentText)]: {\"prompt\" : \"Explain this text\", \"type\": \"comment\"}",
 
@@ -165,8 +170,12 @@ var WORD_FUNCTIONS = {};
 			"showDifference (boolean): whether to show the difference between the original and new text, or just replace it",
 			"type (string): which part of the text to be rewritten (e.g., 'sentence' or 'paragraph')"
 		];
+		func.description = "Use this function when you asked to rewrite or replace some text. If text or paragraph number is not specified assume that we are working with the current paragraph.";
 
 		func.examples = [
+			"if you need to rewrite, respond with:\n" +
+			"[functionCalling (rewriteText)]: {\"prompt\": \"Rewrite\", \"type\" : \"paragraph\"}",
+
 			"If you need to rephrase current sentence, respond with:\n" +
 			"[functionCalling (rewriteText)]: {\"prompt\": \"rephrase sentence\", \"type\" : \"sentence\"}",
 
@@ -371,6 +380,112 @@ var WORD_FUNCTIONS = {};
 
 		return func;
 	}
+	WORD_FUNCTIONS.checkSpelling = function() 
+	{
+		let func = new RegisteredFunction();
+		func.name = "checkSpelling";
+		func.params = [
+		];
+
+		func.description = "Use this function if you asked to check spelling for current paragraph or fix other type of text errors in the current paragraph."
+
+		func.examples = [
+			"if you need to check spelling for the current paragraph, respond with:\n" +
+			"[functionCalling (checkSpelling)]: {}"
+		];
+		
+		func.call = async function(params) {
+
+			let text = await Asc.Editor.callCommand(function(){
+				let par = Api.GetDocument().GetCurrentParagraph();
+				if (!par)
+					return "";
+				par.Select();
+				return par.GetText();
+			});
+
+			let argPromt = "Check spelling and grammar for text:" + ":\n" + text + "\n Answer with only the new corrected text, no need of any explanations.";
+
+			let isTrackChanges = await Asc.Editor.callCommand(function(){
+				let isOn = Api.GetDocument().IsTrackRevisions();
+				if (isOn)
+					Api.GetDocument().SetTrackRevisions(false);
+				return isOn;
+			});
+
+			let requestEngine = AI.Request.create(AI.ActionType.Chat);
+			if (!requestEngine)
+				return;
+
+			await Asc.Editor.callMethod("StartAction", ["GroupActions"]);
+
+			await Asc.Editor.callMethod("StartAction", ["Block", "AI (" + requestEngine.modelUI.name + ")"]);
+
+			let isSendedEndLongAction = false;
+			async function checkEndAction() {
+				if (!isSendedEndLongAction) {
+					await Asc.Editor.callMethod("EndAction", ["Block", "AI (" + requestEngine.modelUI.name + ")"]);
+					isSendedEndLongAction = true
+				}
+			}
+
+			let resultText = "";
+
+			let result = await requestEngine.chatRequest(argPromt, false, async function(data) {
+				if (!data)
+					return;
+				await checkEndAction();
+
+				resultText += data;
+
+				await Asc.Editor.callMethod("EndAction", ["GroupActions", "", "cancel"]);
+				await Asc.Editor.callMethod("StartAction", ["GroupActions"]);
+
+				Asc.scope.text = resultText;
+				await Asc.Editor.callCommand(function(){
+					let par = Api.GetDocument().GetCurrentParagraph();
+					if (!par)
+						return "";
+					par.Select();					
+					Api.ReplaceTextSmart([Asc.scope.text]);
+				});
+			});
+
+			await checkEndAction();
+
+			await Asc.Editor.callMethod("EndAction", ["GroupActions", "", "cancel"]);
+			await Asc.Editor.callMethod("StartAction", ["GroupActions"]);
+
+			Asc.scope.modelName = requestEngine.modelUI.name;
+			await Asc.Editor.callCommand(function(){
+				return Api.GetDocument().SetAssistantTrackRevisions(true, Asc.scope.modelName);
+			});
+
+			Asc.scope.text = resultText;
+			await Asc.Editor.callCommand(function(){
+				let par = Api.GetDocument().GetCurrentParagraph();
+				if (!par)
+					return "";
+				par.Select();
+				Api.ReplaceTextSmart([Asc.scope.text]);
+			});
+
+			await Asc.Editor.callCommand(function(){
+				return Api.GetDocument().SetAssistantTrackRevisions(false);
+			});
+
+			if (isTrackChanges)
+			{
+				await Asc.Editor.callCommand(function(){
+					Api.GetDocument().SetTrackRevisions(true);
+				});
+			}
+
+			await Asc.Editor.callMethod("EndAction", ["GroupActions"]);
+		};
+
+		return func;
+	}
 })();
 
 function getWordFunctions() {
@@ -411,6 +526,7 @@ function getWordFunctions() {
 	funcs.push(WORD_FUNCTIONS.commentText());
 	funcs.push(WORD_FUNCTIONS.rewriteText());
 	funcs.push(WORD_FUNCTIONS.insertPage());
+	funcs.push(WORD_FUNCTIONS.checkSpelling());
 
 	return funcs;